Conexiones arregladas de user y branch
